server/requirements.txt
# Core
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
psycopg2-binary==2.9.9
psycopg2-pool==1.1

# Web Server
gunicorn==21.2.0

# API
flask-cors==4.0.0
python-dotenv==1.0.0

# Development
pylint==3.0.2
black==23.11.0
pytest==7.4.2